huhu

+++ petcare-be/src/main/java/org/example/petcarebe/dto/AppointmentResponse.java
@@ -6,9 +6,9 @@ public class AppointmentResponse {
     private String phone;
     private String date;
     private String time;
-    private float paidAmount;
+    private double paidAmount;
     private double totalAmount;
-    private float depositAmount;
+    private double depositAmount;
     private int petCount;
     private String status;
     private String message;
@@ -22,7 +22,7 @@ public class AppointmentResponse {
 
     // Constructor for getPendingAppointments
     public AppointmentResponse(Long appointmentId, String customerName, String phone, String date, String time,
-                               float paidAmount, double totalAmount, float depositAmount, int petCount) {
+                             double paidAmount, double totalAmount, double depositAmount, int petCount) {
         this.appointmentId = appointmentId;
         this.customerName = customerName;
         this.phone = phone;
@@ -45,12 +45,12 @@ public class AppointmentResponse {
     public void setDate(String date) { this.date = date; }
     public String getTime() { return time; }
     public void setTime(String time) { this.time = time; }
-    public float getPaidAmount() { return paidAmount; }
-    public void setPaidAmount(float paidAmount) { this.paidAmount = paidAmount; }
+    public double getPaidAmount() { return paidAmount; }
+    public void setPaidAmount(double paidAmount) { this.paidAmount = paidAmount; }
     public double getTotalAmount() { return totalAmount; }
     public void setTotalAmount(double totalAmount) { this.totalAmount = totalAmount; }
-    public float getDepositAmount() { return depositAmount; }
-    public void setDepositAmount(float depositAmount) { this.depositAmount = depositAmount; }
+    public double getDepositAmount() { return depositAmount; }
+    public void setDepositAmount(double depositAmount) { this.depositAmount = depositAmount; }
     public int getPetCount() { return petCount; }
     public void setPetCount(int petCount) { this.petCount = petCount; }
     public String getStatus() { return status; }
